feat: add Sidebar component for navigation
Introduce a new Sidebar component to enhance navigation 
throughout the application. The Sidebar includes links to 
various sections as Home, Patterns, Models, Contexts, 
and Chat, each accompanied by an appropriate icon. 
Remove the previous implementation from the 
Header component to streamline the layout and improve 
user experience. Adjust the layout structure to integrate 
the Sidebar effectively.
